// webui/src/generated/api.d.ts
export interface components {
    schemas: {
// ...
        JobResponse: {
            /** Error Code */
            error_code?: string | null;
            /** Result */
            result?: {
                [key: string]: unknown;
            } | null;
            /**
             * Status
             * @enum {string}
             */
            status: "queued" | "running" | "succeeded" | "failed" | "canceled";
        };
// ...
        ProjectResponse: {
            charset: components["schemas"]["CharsetInfo"];
            counts: components["schemas"]["Counts"];
            /**
             * Expires At
             * Format: date-time
             */
            expires_at: string;
            /** Family Name */
            family_name: string;
            /** Name */
            name: string;
            /**
             * Project Id
             * Format: uuid
             */
            project_id: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}